added dropdown

+++ day34__/day34l/src/app/components/employee/list-employee/list-employee.component.ts
@@ -2,6 +2,7 @@ import { Component, inject, OnInit } from '@angular/core';
 import { EmployeeService } from '../../../service/employee.service';
 import { Employee } from '../../../model/models';
 import { HttpErrorResponse } from '@angular/common/http';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-list-employee',
@@ -13,6 +14,7 @@ export class ListEmployeeComponent implements OnInit{
 
 
   employeeService = inject(EmployeeService);
+  private router = inject(Router);
 
   employees:Employee[] =[]
   ngOnInit(): void {
@@ -39,13 +41,24 @@ export class ListEmployeeComponent implements OnInit{
     })
   }
 
-  deleteEmployeeById(id:number | undefined) {
-    this.employeeService.deleteEmployeeById(id).subscribe({
-      next: () => {
-        this.getEmployees()
-      }
-    })
+  deleteEmployeeById(id:number) {
+    if (confirm("do you wanna delete?")) {
+      this.employeeService.deleteEmployeeById(id).subscribe({
+        next: () => {
+          this.getEmployees()
+        }
+      })
+    }
+    
 
+  }
+
+  updateEmployeeById(id:number) {
+    this.router.navigate(['employeeupdate',id])
+  }
+
+  getDetails(id:number) {
+    this.router.navigate(['employeedetails',id])
   }
   
 }
